Refactor ProfileController into separate modules
