+++ src/main/java/com/example/echo_api/controller/profile/ProfileControllerAdvice.java
@@ -7,7 +7,10 @@ import com.example.echo_api.exception.AbstractControllerAdvice;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
-@ControllerAdvice(assignableTypes = ProfileController.class)
+@ControllerAdvice(assignableTypes = {
+        ProfileViewController.class,
+        ProfileInteractionController.class,
+        ProfileManagementController.class })
 public class ProfileControllerAdvice extends AbstractControllerAdvice {
 
 }
